requirements.txt
django==4.2.6
psycopg2-binary==2.9.9
pillow==10.0.1
django-cors-headers==4.2.0
python-decouple==3.8
django-phonenumber-field[phonenumbers]==7.2.0
django-filter==23.3
djangorestframework==3.14.0
drf_spectacular==0.26.5